Refactor learning resource chooser location
